# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VideoForge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/concepts:
    get:
      operationId: getConcepts
      tags: [concepts]
      summary: Get concepts for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Concepts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Concept"
    post:
      operationId: generateConcepts
      tags: [concepts]
      summary: Generate 3 video concepts from analysis
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateConceptsInput"
      responses:
        "200":
          description: Generated concepts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Concept"
  /projects/{projectId}/concepts/{conceptId}/select:
    post:
      operationId: selectConcept
      tags: [concepts]
      summary: Select a concept for the project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: conceptId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{projectId}/media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List media assets for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaAsset"
  /projects/{projectId}/media/{mediaId}:
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Remove a media asset from project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/media/stock:
    post:
      operationId: addStockMedia
      tags: [media]
      summary: Add stock footage to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMediaInput"
      responses:
        "201":
          description: Added media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
  /search/stock:
    get:
      operationId: searchStock
      tags: [media]
      summary: Search Pexels for stock footage
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: perPage
          in: query
          schema:
            type: integer
            default: 15
        - name: type
          in: query
          schema:
            type: string
            enum: [video, photo]
            default: video
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockSearchResult"
...
  /projects/{projectId}/captions:
    get:
      operationId: getCaptions
      tags: [captions]
      summary: Get captions for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Captions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Captions"
    post:
      operationId: generateCaptions
      tags: [captions]
      summary: Generate captions from script/voiceover
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated captions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Captions"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, analyzing, concepts, scripting, media, voiceover, rendering, ready, published]
        referenceUrl:
          type: ["string", "null"]
        referenceTopic:
          type: ["string", "null"]
        selectedConceptId:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        renderProgress:
          type: ["integer", "null"]
        renderStatus:
          type: ["string", "null"]
        outputFormats:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - status
        - createdAt
        - updatedAt
...
    AnalysisResult:
      type: object
      properties:
        projectId:
          type: string
        hookType:
          type: string
        pacing:
          type: string
        storyStructure:
          type: string
        captionDensity:
          type: string
        ctaTiming:
          type: string
        audienceAngle:
          type: string
        estimatedDuration:
          type: ["integer", "null"]
        keyPatterns:
          type: array
          items:
            type: string
        contentWarnings:
          type: array
          items:
            type: string
        summary:
          type: string
        isMock:
          type: boolean
      required:
        - projectId
        - hookType
        - pacing
        - storyStructure
        - captionDensity
        - ctaTiming
        - audienceAngle
        - keyPatterns
        - contentWarnings
        - summary
    Concept:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        angle:
          type: string
        hook:
          type: string
        structure:
          type: string
        targetAudience:
          type: string
        estimatedDuration:
          type: integer
        uniqueElements:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - title
        - angle
        - hook
        - structure
        - targetAudience
        - estimatedDuration
        - uniqueElements
        - createdAt
    GenerateConceptsInput:
      type: object
      properties:
        analysisResult:
          $ref: "#/components/schemas/AnalysisResult"
      required:
        - analysisResult
...
    MediaAsset:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        type:
          type: string
          enum: [video, image, audio, stock_video, stock_photo]
        source:
          type: string
          enum: [uploaded, pexels, generated]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        filename:
          type: string
        duration:
          type: ["integer", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        attribution:
          type: ["string", "null"]
        license:
          type: ["string", "null"]
        createdAt:
          type: string
        orderIndex:
          type: integer
      required:
        - id
        - projectId
        - type
        - source
        - url
        - filename
        - createdAt
        - orderIndex
    StockMediaInput:
      type: object
      properties:
        externalId:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: string
        type:
          type: string
          enum: [stock_video, stock_photo]
        attribution:
          type: string
        license:
          type: string
        duration:
          type: integer
        width:
          type: integer
        height:
          type: integer
      required:
        - externalId
        - url
        - type
        - attribution
    StockSearchResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/StockItem"
        totalResults:
          type: integer
        page:
          type: integer
        perPage:
          type: integer
        isMock:
          type: boolean
      required:
        - results
        - totalResults
        - page
        - perPage
    StockItem:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: string
        type:
          type: string
        duration:
          type: ["integer", "null"]
        width:
          type: integer
        height:
          type: integer
        photographer:
          type: string
        attribution:
          type: string
        license:
          type: string
      required:
        - id
        - url
        - thumbnailUrl
        - type
        - width
        - height
        - photographer
        - attribution
        - license
...
    Captions:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        srtContent:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/CaptionEntry"
        language:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - srtContent
        - entries
        - language
        - createdAt
    CaptionEntry:
      type: object
      properties:
        index:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        text:
          type: string
      required:
        - index
        - startTime
        - endTime
        - text
...