# lib/api-zod/src/generated/types/captions.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * VideoForge API
 * OpenAPI spec version: 0.1.0
 */
import type { CaptionEntry } from './captionEntry';

export interface Captions {
  id: string;
  projectId: string;
  srtContent: string;
  entries: CaptionEntry[];
  language: string;
  createdAt: string;
}